split the instrumenter to share code between node and the browser
