+++ packages/dd-trace/src/platform/browser/loader.js
@@ -1,0 +1,11 @@
+'use strict'
+
+class Loader {
+  reload () {}
+
+  getModules (instrumentation) {
+    return [window[instrumentation.name]]
+  }
+}
+
+module.exports = Loader
